frames: Overhaul DataType to add a get_value(), and use it to make getting values from stuff cleaner.
The best example of how this is useful,
let's say we have a Sequence A containing "foo" as Sequence B,
and Sequence B contains an integer "bar".

Then we can do: seq_A_value.foo
and get back a DataType (Sequence B)
but we can also do: foo.bar
and get back a bare integer, not a datatype.UnsignedLong or whatever.